Assign all posts to the current user

+++ app/Http/Controllers/Admin/PostController.php
@@ -8,6 +8,7 @@ use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Validation\Rule;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 use function GuzzleHttp\Promise\all;
@@ -23,6 +24,8 @@ class PostController extends Controller
     {
         //
         $posts = Post::all();
+        //$posts = Auth::user()->Post::all();
+        $posts = Auth::user()->posts()->orderByDesc('id')->paginate(10);
         return view('admin.posts.index', compact('posts'));
     }
 
@@ -56,6 +59,7 @@ class PostController extends Controller
             //'tags' => 'exists:tags,id',
         ]);
         $validated['slug'] = Str::slug($validated['title']);
+        $validated['user_id'] = Auth::id();
         $_post = Post::create($validated);
         if ($request->has('tags')) {
             $request->validate([
@@ -63,7 +67,7 @@ class PostController extends Controller
             ]);
             $_post->tags()->attach($request->tags);
         }
-        return redirect()->route('admin.posts.index', $_post);
+        return redirect()->route('admin.posts.index');
     }
 
     /**
@@ -84,11 +88,18 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Post $post, Tag $tags)
+    public function edit(Post $post)
     {
         //
+        $categories = Category::all();
         $tags = Tag::all();
-        return view('admin.posts.edit', compact('post', 'tags'));
+        /* condizione per limitare l'accesso ad altri utetnti */
+        if (Auth::id() === $post->user_id) {
+
+            return view('admin.posts.edit', compact('post', 'tags', 'categories'));
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -101,22 +112,26 @@ class PostController extends Controller
     public function update(Request $request, Post $post)
     {
         //
-        $validated = $request->validate([
-            //'title' => ['required', Rule::unique('posts')->ignore($post->id)],
-            'title' => 'required',
-            'image' => 'nullable',
-            'sub_title' => 'nullable',
-            'content' => 'nullable',
-            'category_id' => 'nullable | exists:categories.id',
-        ]);
-        if ($request->has('tags')) {
-            $request->validate([
-                'tags' => ['nullable', 'exists:tags,id']
+        if (Auth::id() === $post->iser_id) {
+            $validated = $request->validate([
+                //'title' => ['required', Rule::unique('posts')->ignore($post->id)],
+                'title' => 'required',
+                'image' => 'nullable',
+                'sub_title' => 'nullable',
+                'content' => 'nullable',
+                'category_id' => 'nullable | exists:categories,id',
             ]);
-            $post->update($validated);
-            $post->tags()->sync($request->tags);
+            if ($request->has('tags')) {
+                $request->validate([
+                    'tags' => ['nullable', 'exists:tags,id']
+                ]);
+                $post->update($validated);
+                $post->tags()->sync($request->tags);
+            }
+            return redirect()->route('admin.posts.index');
+        } else {
+            abort(403);
         }
-        return redirect()->route('admin.posts.index');
     }
     /**
      * Remove the specified resource from storage.
@@ -126,8 +141,12 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        //
-        $post->delete();
-        return redirect()->route('admin.posts.index');
+        if (Auth::id() === $post->user_id) {
+            //
+            $post->delete();
+            return redirect()->route('admin.posts.index');
+        } else {
+            abort(403);
+        }
     }
 }
